# textLayout speed on screen