Part of #539 

This addresses some performance problems with textLayout when showing a lot of text on the screen.  It does not address how to skip non-visible text.

* fixes the bug where performance degrades as you were scrolled lower
* for me this about doubles the speed of processing visible text in Debug mode:
  * for ~24K ascii chars visible, went from 7fps -> 14fps
* ReleaseFast sees less improvement (maybe 55fps->65fps)

What I learned is that performance (at least in Debug) is dominated by:
* looking up GlyphInfo for each codepoint
* looking up kerning for each codepoint pair

For GlyphInfo we are now cheating by prebaking all ascii characters.  I wonder if there is a better way to do this, either by making GlyphInfo smaller, or having a better lookup datastructure.

For kerning we are reusing most of the information from textSize in renderText.  The key here is assuming that non-zero kerning pairs are rare.  Maybe something like a bloom filter would work better, if we could get the kerning table straight from the font.